Generate Unit Test: move integration test to separate file
Same as we have done for Document Code and Custom Commands.
