+++ agent/src/__tests__/generate-unit-test/src/trickyLogic.ts
@@ -1,0 +1,12 @@
+export function trickyLogic(a: number, b: number): number {
+    if (a === 0) {
+        return 1
+    }
+    if (b === 2) {
+        return 1
+    }
+
+    return a - b
+}
+
+/* CURSOR */
